2009.
Nadal wins his first Australian Open. Federer finally wins the French Open!
Nadal disappointedly failed to defend his Wimbledon title, so made a note there
to highlight this. del Potro wins the US Open in epic fashion in a time when
Federer still looks to be on the top of his game.